requirements.txt
fastapi~=0.75.0
uvicorn[standard]
pytest
requests

pydantic~=1.10.22
psycopg2-binary~=2.9.10